# openapi.yaml
openapi: 3.0.3
info:
  title: GoShort API
  version: 1.0.0
  description: |
    GoShort 后端 API 文档（基于当前代码实现生成），用于前端代码生成。

    - API 前缀: `/api/v1`
    - 认证方式: `Authorization: Bearer <JWT>`
    - 返回风格: 统一包含 `success`，业务错误多数包含 `code/message`
servers:
  - url: https://localhost
    description: Nginx HTTPS 入口
  - url: http://localhost:8081
    description: 直连 api-service（开发调试）
...
paths:
...
  /api/v1/links:
    post:
      tags: [Links]
      summary: 创建短链接
      operationId: createLink
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLinkRequest'
      responses:
        '200':
          description: 创建成功或链接已存在（同样返回 200）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '400':
          description: 参数错误或短码冲突
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/MiddlewareErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedMiddleware'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Links]
      summary: 获取当前用户链接列表
      operationId: getLinks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLinksResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedMiddleware'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/links/GetLinksByAlias:
    get:
      tags: [Links]
      summary: 按别名筛选当前用户链接
      operationId: getLinksByAlias
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: alias
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLinksResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedMiddleware'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/links/{id}:
    delete:
      tags: [Links]
      summary: 删除短链接
      operationId: deleteLink
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedMiddleware'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/MiddlewareErrorResponse'
        '404':
          description: 链接不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 请求无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: BAD_REQUEST
            message: 请求无效
    InternalError:
      description: 系统错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: INTERNAL_ERROR
            message: 系统错误
    UnauthorizedMiddleware:
      description: 未认证（中间件直接返回）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MiddlewareErrorResponse'
          examples:
            missingAuth:
              value:
                error: Authorization header required
            badFormat:
              value:
                error: Invalid authorization header format
            expired:
              value:
                error: Invalid or expired token
...
  schemas:
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
        message:
          type: string
        details:
          type: string
      required: [success, code, message]
    MiddlewareErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    CreateLinkRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
        alias:
          type: string
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: boolean
          nullable: true
        short_code:
          type: string
          nullable: true
      required: [url]
    LinkResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            link_id:
              type: integer
              format: int64
            short_code:
              type: string
            original_url:
              type: string
            short_url:
              type: string
            alias:
              type: string
            is_active:
              type: boolean
              nullable: true
            expires_at:
              type: string
              format: date-time
              nullable: true
            created_at:
              type: string
              format: date-time
    ListLinksResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            links:
              type: array
              items:
                $ref: '#/components/schemas/LinkResponse'
            total:
              type: integer
              format: int64
            page:
              type: integer
            limit:
              type: integer
          required: [links, total, page, limit]
...